Fixed Clang Tidy issues; resolved infinite recursion error with struct reference lookup

+++ packages/compiler/src/ast/nodes/package.cpp
@@ -21,7 +21,7 @@ std::unique_ptr<AstPackage> stride::ast::parse_package_declaration(
     TokenSet& set
 )
 {
-    const int initial_offset = set.position();
+    const size_t initial_offset = set.position();
     const auto reference_token = set.expect(TokenType::KEYWORD_PACKAGE);
     const auto package_name = set.expect(TokenType::IDENTIFIER, "Expected package name").lexeme;
